String-style units, assertions on make-quantity
Switching to string-style units for better clarity. At some point we'll
parse and check these for valid dimensionality &/or canonical form.

(Not doing any sort of keyword conversion of the units... yet. I'm
worried that eg.  `(keyword "m**2/year")` will run afoul of keyword
namespacing and have to think it through a bit).

On consideration, feeding make-quantity the wrong things is an assert/error,
not an exception that we might ever recover from at runtime.