Stub in interface
Make a naive implementation that is just enough to test the construction
of record-type quantities with a partial in a separate project.
Currently this has zero protection against unregistered units, or
implementation of arithmetic or conversion.

Going down a record-based implementation to experiment with getting the
faster dispatch (that ought to be possible) over the older vector-based
tuple implementation.